Trying to make it so that there is one page that shows register and login based on user button click
